+++ ios/BlurUtils.h
@@ -4,5 +4,6 @@
 
 + (NSNumber *)clipRadius:(NSNumber *)radius;
 + (UIBlurEffectStyle)blurEffectStyle:(NSString *)style;
++ (UIColor *)colorFromString:(NSString *)colorString;
 
 @end