+++ ios/VibrancyView.h
@@ -13,8 +13,10 @@ NS_ASSUME_NONNULL_BEGIN
 #if !TARGET_OS_TV
 
 @property(nonatomic, copy, nullable) NSString *overlayColor;
+@property(nonatomic, copy, nullable) NSString *reducedTransparencyFallbackColor;
 @property(nonatomic, copy, nullable) NSNumber *blurRadius;
 
+@property(nonatomic, strong, nullable) UIView *reducedTransparencyFallbackView;
 @property(nonatomic, strong, nullable) BlurViewEffect *blurEffect;
 @property(nonatomic, strong, nullable) UIVisualEffectView *blurEffectView;
 @property(nonatomic, strong, nullable) UIVibrancyEffect *vibrancyEffect;