+++ ios/VibrancyViewManager.mm
@@ -19,6 +19,7 @@ RCT_EXPORT_MODULE(VibrancyView);
 }
 
 RCT_EXPORT_VIEW_PROPERTY(overlayColor, NSString);
+RCT_EXPORT_VIEW_PROPERTY(reducedTransparencyFallbackColor, NSString);
 RCT_EXPORT_VIEW_PROPERTY(blurRadius, NSNumber);
 
 @end