+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# [1.4.0](https://github.com/DanielAraldi/react-native-blur-view/compare/v1.3.3...v1.4.0) (2026-03-03)
+
+
+### Bug Fixes
+
+* **android:** only declare `setReducedTransparencyFallbackColor` member method generated ([af418a9](https://github.com/DanielAraldi/react-native-blur-view/commit/af418a98e9a3e25d49775b967b5fe404ff5ddfcf))
+
+
+### Features
+
+* **ios:** add the new `reducedTransparencyFallbackColor` property ([289acd5](https://github.com/DanielAraldi/react-native-blur-view/commit/289acd5ae623fbf267211d9479b13f65ce38a114))
+
+## [1.3.3](https://github.com/DanielAraldi/react-native-blur-view/compare/v1.3.2...v1.3.3) (2026-03-02)
+
+
+### Bug Fixes
+
+* **android:** add new `downscaleFactor` property for `BlurView` component (Android only) ([2da16e5](https://github.com/DanielAraldi/react-native-blur-view/commit/2da16e561140b0f87b2b0c6867e449942b30fe0f))
+
+## [1.3.2](https://github.com/DanielAraldi/react-native-blur-view/compare/v1.3.1...v1.3.2) (2026-02-27)
+
+
+### Bug Fixes
+
+* filled adjustment of the overlay color ([a67cc12](https://github.com/DanielAraldi/react-native-blur-view/commit/a67cc1282793c2750b013c7e9a2dbc19814bc534))
+
 ## [1.3.1](https://github.com/DanielAraldi/react-native-blur-view/compare/v1.3.0...v1.3.1) (2025-12-19)